+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.0.*
 
+### Current
+ * jQuery adapter now will JSON stringify data and add `application/json` Content-Type when data is present. If data is already a string it will be left alone.
+
 ### 0.1.0
  * Changed `start` to `connect` which now returns a promise
  * Changed onReady/onReject callbacks to simple `on` event handlers